Add files via upload
In the last decade, the growing influence of social medias on the political sphere has opened an interesting field for predicting election results. This is achieved by computing a popularity index of election candidates, based on sentiment analysis in data collected from Twitter. Our paper contributes to this literature by proposing an improvement of the existing indices. This improvement consists in penalizing the indices by a potential subjectivity that can be observed in users’ comments. We then apply these penalized indices to predict the 2018 DRC presidential election results using ARIMA modeling, and compare their accuracies to those of non-penalized indices. The results show that penalization substantially increases the accuracy metrics in predicting candidates popularity.